Funding acknowledgement; missing licence
Hi,

Is this repository still relevant or is it entirely outdated? Please add this to the readme file.
The funding acknowledgement is necessary in any case. 
(The licence is relevant if you want the code to be reused. Code without a licence should not be included in any other projects.)